Move RoadStationPlacementArgs to CreateRoadStation.h
