+++ src/OpenLoco/src/Windows/Construction/StationTab.cpp
@@ -2,6 +2,7 @@
 #include "Construction.h"
 #include "Drawing/SoftwareDrawingEngine.h"
 #include "GameCommands/GameCommands.h"
+#include "GameCommands/Road/CreateRoadStation.h"
 #include "GameCommands/Track/CreateTrainStation.h"
 #include "Graphics/ImageIds.h"
 #include "Input.h"
